Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    int x;
-    scanf("%d",&x);
+    float x;
+    scanf("%f",&x);
     float Area = 3.14*x*x;
     printf("Area: %.2f",Area);
     return 0;
